Move collections component js to collections directory
 🐿 v2.9.0
